introduced AlignCommand
