+++ src/test/java/frc/robot/commands/AlignCommandTest.java
@@ -10,7 +10,7 @@ import edu.wpi.first.hal.HAL;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 
-public class DriveCommandTest {
+public class AlignCommandTest {
 
 	@BeforeAll
 	static void setup() {
@@ -25,13 +25,13 @@ public class DriveCommandTest {
 	void testAngularDisplacement() {
 		assertEquals(
 				45.0,
-				DriveCommand.angularDisplacement(
+				AlignCommand.angularDisplacement(
 						new Pose2d(0, 0, Rotation2d.fromDegrees(45)), new Pose2d(0, 1, Rotation2d.fromDegrees(0)))
 						.getDegrees(),
 				0.1);
 		assertEquals(
 				135.0,
-				DriveCommand.angularDisplacement(
+				AlignCommand.angularDisplacement(
 						new Pose2d(0, 0, Rotation2d.fromDegrees(-45)), new Pose2d(0, 1, Rotation2d.fromDegrees(0)))
 						.getDegrees(),
 				0.1);
